+++ src/__internal__/input/input-presentation.test.tsx
@@ -4,6 +4,7 @@ import userEvent from "@testing-library/user-event";
 import InputPresentation from "./input-presentation.component";
 import Input from "./input.component";
 import { InputContext, InputGroupContext } from "../input-behaviour";
+import StyledSelect from "../../components/select/select.style";
 
 test("renders presentational div and presentation container elements", () => {
   render(<InputPresentation>Children</InputPresentation>);
@@ -106,8 +107,27 @@ test("triggers a passed function via the `onMouseLeave` prop from the input cont
   expect(onMouseLeaveMock).toHaveBeenCalled();
 });
 
-/* Styling test for coverage */
+test("applies the correct z-index values when inside an open select", () => {
+  render(
+    <StyledSelect isOpen>
+      <InputPresentation>
+        <Input />
+      </InputPresentation>
+      ,
+    </StyledSelect>,
+  );
 
+  const inputPresentation = screen.getByRole("presentation");
+  const backdropIndex =
+    getComputedStyle(inputPresentation).getPropertyValue("z-index");
+
+  // non-default value
+  expect(backdropIndex).toContain("--adaptiveSidebarModalBackdrop");
+  // default value
+  expect(backdropIndex).toContain("9999");
+});
+
+/* Styling test for coverage */
 test.each(["left", "right"])(
   "when `hasIcon` is true and the `align` prop is %s, the corresponding padding attribute is 0",
   (alignValue) => {